IMAP - replace existing IMAP James backend with a new version based on tagged searches
- Delete the old IMAP database and migrate the deletion
- Heavily modify the James integration layer to operate on saved
  searched tagged as being IMAP searches.
- Fix James termination
  - James has two cases of not properly shutting down it's threaded
    objects.
    - First is the HashedWheelTimer in the ChannelPipelineFactory in
      IMAPServer.  It was not being shut down properly becaues the
      ChannelPipelineFactory dosen't implement
      ExternalResourceReleasable.  The solution is to subclass and
      implement that interface, shutting down the HashedWheelTimer.
    - The second is the IdleProcessor in the DefaultProcessorChain.  If
      it's using it's keepalive timer, than it creates a
      ThreadPoolExecutor that never is terminated.  There's no API to
      allow it to be terminated, so the only solution at this point is
      to disable the keepalive timer.
- Allow search folders to be updated
  - Make EnkiveImapMailbox be a singleton per session
  - Add a timer to it that updates the search on a configured interval
  - Send a notification to a client IDLEing on the mailbox when it
    updates

+++ src/java/com/linuxbox/enkive/message/search/AbstractMessageSearchService.java
@@ -112,8 +112,8 @@ public abstract class AbstractMessageSearchService implements
 				// This re-ran the entire search, as it was a migrated search.  We need to
 				// wipe out all previous results or we'll get duplicates
 				query.invalidateUID();
-				result.setNextUID(0);
-				result.setMessageIds(new HashMap<Integer, String>());
+				result.setNextUID((long)0);
+				result.setMessageIds(new HashMap<Long, String>());
 			}
 			result.addMessageIds(resultMessageIDs.values());
 			query.setLastMonotonic(resultMessageIDs.lastKey());
